requirements.txt


openai==0.28
pymongo==4.6.1
python-dotenv==1.0.0
googlemaps==4.10.0
streamlit==1.29.0
langchain==0.1.0
langchain-openai==0.0.2
requests==2.31.0
pydantic==2.5.0
typing-extensions==4.8.0
dataclasses==0.8
enum34==1.1.10
